app/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Jinja2==3.1.2
mysqlclient==2.1.1
paramiko==2.11.0
pysftp==0.2.9
python-dotenv==0.21.0
